working on the store, got products broken into components
