+++ client/src/index.css
@@ -7,3 +7,7 @@
 * {
     font-family: 'Roboto', sans-serif;
 }
+
+body {
+    background-color: rgb(247, 235, 216);
+}
