--- client/src/userFront/component/CheckoutBtn.jsx
@@ -1,15 +1,0 @@
-import * as React from 'react';
-import Button from '@mui/material/Button';
-import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
-import Stack from '@mui/material/Stack';
-
-export default function IconLabelButtons() {
-  return (
-    <Stack direction="row" spacing={2}>
-      <Button variant="outlined" startIcon={<ShoppingCartIcon />}>
-        ADD TO CART
-      </Button>
-    </Stack>
-  );
-}
-
