--- client/src/userFront/component/cartContext.js
@@ -1,23 +1,0 @@
-import React, { createContext, useContext, useState } from 'react';
-
-const CartContext = createContext();
-
-export const useCart = () => {
-    return useContext(CartContext);
-}
-
-export const CartProvider = ({ children }) => {
-    const [cart, setCart] = useState([]);
-
-    const addToCart = (item) => {
-        setCart(prevCart => [...prevCart, item]);
-    }
-
-    return (
-        <CartContext.Provider value={{ cart, addToCart }}>
-            {children}
-        </CartContext.Provider>
-    );
-}
-
-
