+++ client/src/userFront/component/shop/Cart.js
@@ -1,0 +1,18 @@
+import { useCart } from './cartContext';
+
+function Cart() {
+    const { cart } = useCart();
+
+    return (
+        <div>
+            <h2>Your Cart</h2>
+            {cart.map(item => (
+                <div key={item.id}>
+                    <span>{item.name} - {item.size} - {item.color} - Quantity: {item.quantity}</span>
+                </div>
+            ))}
+        </div>
+    );
+}
+
+export default Cart;
