+++ client/src/userFront/component/shop/Colors.jsx
@@ -1,0 +1,16 @@
+import * as React from 'react';
+import Button from '@mui/material/Button';
+import Stack from '@mui/material/Stack';
+
+export default function ColorButtons({ label }) {
+  return (
+    <div>
+    <Stack direction="row" spacing={2}>
+      <Button variant="outlined">
+      {label}
+      </Button>
+    </Stack>
+    </div>
+  );
+}
+
